ellipsis bug fix (#11)
### Summary
The problem was that invalidateData(), which gets the width of the EditText and uses it to configure the chip width, is only called in the View lifecycle before the layout width of the TextView is set. getWidth() always returned 0. I ended up overriding the onDraw() method and called invalidateData() from there as well, since onDraw() is called in the lifecycle after the View's layout is set.
### Test Plan
In the nachos test app, set a few nachos that should appear as soon as the app launches and check that the chips do not ellipsis. Make sure that everything else works too.
### Reviewer
@simon-tse-hs @neil-power-hs @wesley-alcock-hs 
### JIRA Ticket
https://jira.hootsuitemedia.com/browse/AND-4503