Make it possible to set a specific background corner radius
Currently the background corner radius is based on the chip height `(height / 2)`, so we implement a way to override the default value.

Although the Material Design Chips documentation only displays the chips with the same roundness, there's nothing explicitly documented about it, so I believe it'd be okay for the library to have some flexibility on that regard.